requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.6.0
black==19.10b0
blessed==1.17.4
boto==2.49.0
boto3==1.12.32
botocore==1.14.17
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.8
Django==3.0.3
django-countries==5.5
django-dotenv==1.4.2
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
future==0.16.0
idna==2.7
jmespath==0.9.5
joblib==0.14.1
JPype1==0.7.2
jsonschema==3.2.0
konlpy==0.5.2
lxml==4.5.0
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.0.0
pipreqs==0.4.10
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyNaCl==1.3.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
regex==2020.1.8
requests==2.20.1
requests-oauthlib==1.3.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
semantic-version==2.5.0
six==1.11.0
sklearn==0.0
smart-open==1.9.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==1.6.2
toml==0.10.0
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
yarg==0.1.9
psycopg2-binary==2.8.4
django-storages==1.9.1
sentry-sdk==0.14.3
